SLA report: Default SLA values when none is found
The old reports only contained data for months that had SLA values in
them. This would result in new reports having their SLA values neatly
added to the SLA input fields but for the old reports they would always
start from the first month. Add 0.0 as default for SLA values if none is
found in the db.

This is part of fixing MON-6154.

Signed-off-by: Robin Hagman <robin.hagman@op5.com>