Keep a list of all classes for autoload
Don't do any creative magic of class names for autoloading. It's better
to on build time generate a list of all classes and keep with the
module, so we can once when loading just list all classes in the system
and quickly load a class.

it readuced the number of calls to Kohana::find_file with about a
hundred when running "php index.php tac/index". With 10 or so modules,
that means about a thousand file stats.

Probably much more in a ORM intensive page.

Without numbers, the page feels lot quicker after this patch

Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>