Generate a list of all views in the system
Loading views means a lot of find_files, which means a lot of disc
access. But we already know what files we have, so just load a simple
list of all files in each module, which can be generated at build.

This reduces almost every find_file call that is left. What is left is
however config and widgets, so that method can't be removed.

Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>