form model: Make form fields optional/take callback if value is missing
Manual testing of MON-9384 discovered that we could not set an empty
value for a field, in the specific case, a widget's refresh rate.

We mitigate this by introducing Form_Model::set_missing_fields_cb() and
the MissingValueException. When calling Form_Model::process_data(),
set_missing_fields_cb() is called if MissingValueException is thrown,
looking for special handling of a missing value. See the test cases for
more explicit use cases.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>