livestatus: Inject connection as dependency
It is much easier to mock classes (op5livestatus) if they are upfront
with their requirements (the constructor now wants a
op5livestatus_connection) instead of hiding it (the constructor
previously wanted an array in which a single key was of importance).
This also makes the unit test at bit easier to read.

The reason I modified this right now, is that another issue was failing
with a nonsense message (a php warning because of unchecked indices)
instead of saying "I failed to fulfill a contract".

Also, sorry, but I had to remove a couple of too obvious, and a bit
spammy, comments :)

Signed-off-by: Carl Helmertz <chelmertz@op5.com>