cmd: Friendlier message for users without correct authorization levels
We focus on the error message for the most common use case for when a
command submission is failing this early - insufficient authorization
rights.

There's a lengthy inline comment that explains the contradiction between
the factually correct thing to say, and the helpful thing to say.

We also style the error message with a nice blue box (let's not scare
the users with red, they meant well). The large diff in the view file is
just because I removed a level of indentation.

This fixes MON-9487.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>