Move all unit tests into the same folder
There is no reason to keep the unit tests for the "phplib" (src/op5/*,
the stuff that gets installed into /usr/share/php) separated from the
rest of the Ninja unit tests.

There are at least two large reasons for keeping all unit tests in the
same directory:
 - easier to find all of them, especially for new users
 - the Makefile executed tests in the two folders in sequence. That's
   poor design, since the second line in the Makefile will never run if
   there was an error in the first one. We *always* want all unit tests
   to run

We also move the directory up a bit so that PHP files are directly in
test, instead of test/unit_test/tests. That makes it a lot easier to
type and refer to everywhere.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>