dashboard: Handle if 'position' was not migrated with widget settings
Migrating the widgets from the SQL table ninja_widgets to
dashboards/dashboard_widgets skipped setting a value for the "position"
column if there was no stored order of the widgets.

This caused a bug to appear: when trying to add a new widget, the GUI
responded with "Could not save new widget to settings".

Now we default to placing widgets in the top left if there is missing
data. While checking for missing data, MF and me found this value:

[root@head664 ~]# mysql merlin -e 'select count(*) from
dashboard_widgets where position = ""'
+----------+
| count(*) |
+----------+
|    12184 |
+----------+

This fixes MON-9491

Signed-off-by: Carl Helmertz <chelmertz@op5.com>