Attach previous exception in ORM pools
This helps a lot when debugging test cases that includes the mock
driver. The mock driver is not very extensive yet, so it crashes on
"normal" use cases (such as querying an empty table) with weird errors.

PHPUnit displays the chained exception in a pretty helpful way, so let's
help it help us.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>