kohana: Remove unused $has_error, typehint exception_handler, exit(1)
Even if it is annoying with really small changes, I think it has the
benefit of a) creating good diffs that are easy to git blame; and more
importantly, b) remove ~10 year old code that somebody thought was a
good idea when they wrote Kohana, but isn't used yet. Less code is
easier to read.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>