ninja_db_init: Abort with 1 if php any php script fails
Since ninja_db_init.sh is run as a postlet script (%post) in the spec
file, a bad exit code will not choke all of the installation but it will
at least indicate that something went wrong and also stop executing the
rest of the DB upgrade scripts. Error early, etc.

The recovery may be annoying to do, but at least the user is certain
that the script failed at a specific step. Before this patch, steps 1,
2 and 4 may have 'succeeded', but if 4 depended on the failed step 3,
we're screwed.

The reason for this commit is to discover failures earlier, originally
spurred by a very spammy output log from building Ninja; "Upgrading
ninja db to v12" produced a HTML template with no actual data. Very
annoying.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>