React to the first login of a user and create the default dashboard
This is done by storing the login time of user, if there is no previous login
time then treat it as the first login and run the ninja.first_login event.

The first_login event is handled in the widgets module hook and creates a
dashboard for the user loging in using the default dashboard configuration
available.

This fixes MON-9516

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>