Adding 500 view
This adds a "500 Internal server error" view to Ninja.
At the moment it is needed by Trapper in issue DEVINFRA-498
to handle possible errors when user downloads a LUA script.

The view is accessed directly by the Traps_Controller class,
just like the JSON view is used by some controllers.

Signed-off-by: Joachim Ekstrand <jekstrand@op5.com>