autocompleter: Suggest services if input matches host name
Christian Nilsson was having a hard time when he tried to find a service
with a generic name ("ping") on a somewhat large Monitor installation.
He wants to narrow down the search, and he knows the name of the host
that the service belongs to. When I live patched his system with the
autocomplete.php part of this patch, he felt that this solution could
benefit users even if it doesn't fix MON-9519 completely.

This mitigates part of MON-9519.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>