tac: The Options menu should refer to specified nouns
The options menu had these top level elements:
- Add widget
- Select layout
- Rename
- Delete

The first two menu items are clear in intentions, the latter two not so
much. They are referring to the dashboard that we are currently looking
at but the user might not even have assimilated the notion of dashboards
so this might be a bit confusing.

JIRA acts as a great example, where each menu item has both the verb as
the action, and the noun as the subject of the action.

To make it even a bit clearer, we prefix the menu label with Dashboard,
so that it becomes "Dashboard options".

Signed-off-by: Carl Helmertz <chelmertz@op5.com>