Remove unused execution path to display livestatus exception
The run of system.execute is wrapped in a try catch, the catch is the one place
where we catch the Op5LivestatusException and pass it on to the
application.livestatus event, but this will never happen.  As system.execute
runs only Kohana::instance which catches all exceptions internally.

And since no other execution path runs application.livestatus no other path
will run the error controller with show_livestatus and hence no controler will
load the view livestatus.

Therefor I felt it was safe to remove this path.

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>