view: render(false) is deprecated, __construct() always takes filename
The `View::render(false)` part was already deprecated through regular
logging, now it is also deprecated through the used of
`flag::deprecated()`.

`View::__construct()` used to claim that a filename is optional, and
then later on complain (via an exception) if there was no filename. It's
a no-brainer to make it mandatory.

We also start to throw exceptions when `Kohana::get_view()` fails to
look up a view, because just returning `null` could lead to (unintended
and not pretty) fatal errors later on. Besides, we control all of the
calling locations. This change goes together with the one mentioned in
the previous paragraph.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>