test: Refactor ninja report tests to a phpunit test
This will help us get rid of spam in the installation log, since we wrap
the test output in an output buffer. We also run all of the report tests
instead of exiting at the first failing test.

By rewriting the test as a phpunit test, we only get a single directory
of files to unit test with phpunit. Previously, since both the report
test suite and the second (more general) test suite were separate make
targets, we always execute all of the phpunit tests. That is a really
good thing.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>