report_options: "Last week" represented the date one week too early
Instead of trying to make php's strtotime() work everything out by
giving it a complicated first argument, we adjust the date by giving it
a second argument so that the relativeness is correct.

I stole this solution from Philip Eklöf, whose research is displayed on
the page of the associated issue.

This fixes MON-7264

Signed-off-by: Carl Helmertz <chelmertz@op5.com>