report: Correct thisweek, test last{7,31}days, make tests more robust
Robin Hagman noticed that 'thisweek' would probably not work on a
Sunday, and he was correct! The tests proves me right.

Additionally, I switched out 2016 in all test fixtures to 2015, which
caught some untestable code that needed to be testable.

I added exceptions that are never thrown, but I will remove them in the
following commit, just to make sure that the signature of
`Report_options::calculate_time()` is unchanged. I will actually take
out the meaty parts of it and put it into a helper, since it is unit
testable instead of having to know the inner workings of
`Report_options`.

This fixes MON-9558

Signed-off-by: Carl Helmertz <chelmertz@op5.com>