Overhaul object details page (extinfo)
This commit completely changes the appearance of the object details page, using
new vector/font icons and c3/d3 for charts (gauges). Its general appearance is
updated to better fit with the new dashboard look and feel.

During this development some errors occured with other code that are included
in this fix as it will not build otherwise (see nacoma-urls and
naemonmonitoredobject)

In addition it rewrites monitoring.feature to not actually execute commands but
instead it mocks object states and verifies that they are displayed properly
during that state.

This fixes MON-9356

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>