log: Do not try to chmod() a file that was not created
This caused a build of HTTP API to fail. The warning [1] that was
generated when `chmod()` failed, was appended to the response that was
supposed to be json only. That made some tests to the HTTP API fail.
Here's a part of the output of running the python nose tests against the
HTTP API:

======================================================================
FAIL: command.test_unauthorized_ack_service_problem
----------------------------------------------------------------------
Traceback (most recent call last):
	File "/usr/lib/python2.6/site-packages/nose/case.py", line 182, in runTest
	self.test(*self.arg)
	File "/opt/monitor/op5/test/nose/command.py", line 171, in test_unauthorized_ack_service_problem
	'comment': 1,
	File "/opt/monitor/op5/test/nose/command.py", line 30, in request_json
	ok_(False, "Could not extract JSON (%d):\n%s" % (r.status_code, r.text))
	File "/usr/lib/python2.6/site-packages/nose/tools.py", line 25, in ok_
	assert expr, msg
	AssertionError: Could not extract JSON (403):
{"error":"Forbidden","full_error":"You cannot submit the command ACKNOWLEDGE_SVC_PROBLEM due to insuffecient rights."}
Warning: chmod(): No such file or directory in /usr/share/php/op5/log.php on line 258

Signed-off-by: Carl Helmertz <chelmertz@op5.com>