group downtimes: Yield correct error messages
This patch fixes a bug that would prevent error messages from a
scheduled downtime on a service-, or hostgroup from being propagated and
instead simply show "No hosts in hostgroup".

It also fixes the default error message for the different permutations
to better reflect what was wrong when a group actually was missing a
certain type of object (services, hosts), instead of always talking
about hosts in hostgroups (even for services in servicegroups).

This fixes MON-8661.

Signed-off-by: Anton Lofgren <alofgren@op5.com>