popover: Make adjustments better
Previous to this patch, the popover adjustments were made based on width
and horizontal offset only. They are now made based on vertical offsets
and height as well to make sure that popups rendered near the bottom of
the screen are still visible.

Also, the position adjustments are now done a little bit later to make sure
that we have loaded the resource that we're supposed to adjust against.

In order to make the implementation a bit more consistent, we now also
create paragraphs when a string is passed for popup'ing.

This fixes MON-8721.

Signed-off-by: Anton Lofgren <alofgren@op5.com>