Update interface of dashboards menu
To reflect the UX requirements an update was needed in the Menu model, i.e.
that we need to support separators in the menu, some labeled, some not.

In addition I felt that it was ugly to require specification or order as to not
fall to "undefined" rendering order, now the menu renders in insertion order if
no order parameter has been given to the menu items.

This is part of MON-9553

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>