Kohana: Start rendering the prettified 503 (e.g. ls is down)
Instead of trying to manipulate the controller and running more events
that adds even more code branches to try to check, we can exit early
from a most likely unrecoverable state. (We do need the JS from our
bundles though, so one event is run.)

The lightbox that pops up even mentions "[...] you may attempt to
refresh this page in your browser [...]" so I think we're on the safe
side here.

Just in case we cause Error_Controller::show_503() to throw an
exception, we wrap it so that it at least shows the orange, Kohana,
stack trace, but at this moment in time, that try-catch clause will
never kick in. I don't want to be the optimistic one though, so let's
throw it in there.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>