livestatus: Put more logic within the retry-loop
Just to be sure, let's put all of the "verify the data from livestatus"
logic within the retry-loop. This might catch errors. I say might,
because I haven't reproduced it. It is however really cheap to just say
"let's just try again" if the customer regularly runs into this.

As a bonus, we now log the query in all bad exit points of
`livestatus::query()`.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>