Always return a set from set_by_key
It may be an empty set but it is always nice to be able to expect a single type
returned from this function.

Specialcase StatusPool_Model as the status pool only has one object, that cannot
be filtered, to return self::all() always. So that the generic command
submission works as expected for this object type as well.

If no objects are found given the keys supplied then show a generic information
message.

This fixes MON-9582

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>