auth: Add perms + tests for migration scripts
When fixing the "share dashboards", we added a new auth right. Since
nobody wanted to write tests for that before, I found some nasty things
which are corrected now as well. Now the following things are checked:

- all auth rights in `op5Authorization::get_all_auth_levels()` are
  validated to exist in the shipped "admins" group, in
  etc/auth_groups.yml
- all auth rules extracted from the "acl" in `user_mayi_authorization`
  are now validated to exist, by comparing them with the result of
  `op5Authorization::get_all_auth_levels()`
- the migration script is now tested, and fails if we did not properly
  modify the "admins" group in etc/auth_groups.yml

These are some really nice additions that are added as part of fixing
MON-9539.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>