acl: Make !dashboard_share work as intended
I had missed (the obvious signs of) the fact that it is only the first
matching rule that decides whether or not the action is denied. We loop
through all of the ACL lines, but that is only to find any denying rules
(even if we alredy approved the auth right).

Signed-off-by: Carl Helmertz <chelmertz@op5.com>