Rename quarks' props to foreign_key/foreign_table
This also implies quite a large refactoring from the `Dashboard_Model`'s
perspective: the explicit `share_with()` family of methods are removed
and `set_read_perm()`/`get_read_perm()` now return arrays of tables and
keys instead of an array of permission quarks. This makes the frontend
unaware of permission quarks, which is a good thing.

We also move `get_permission_quark()` to the `Object_Model` so that all
objects may be associated with rights.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>