mayi: Prettify auth rights/mayi rights in the ACL
The logic for attaching all dashboard related items to the main menu is
moved out of the hook into a helper, because it is so annoying to test
events in an encapsulated way.

The test I wrote actually worked for the last version of the ACL as
well. It looks much prettier now though, especially without negotiating
an auth right, which can be confusing to read, even for me who wrote it.

This is a fixup for MON-9539

Lastly, this patch happened to get some janitoring stuff in it as well,
because I came upon it while writing the rest of the patch.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>