lightbox: Hovering the "close"/"X" button makes the pointer a hand
There was a recent change in the CSS so that <a> elements now require a
href attribute in order to switch out the mouse pointer to become a
hand. This is of course layman speak for the change a { cursor: pointer;
}, to a[href] { cursor: pointer; }.

Since we add a href attribute, we need to stop the event from going
through as well.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>