test: Remove deprecated/commented out variables from nagios.cfg.in
The variables max_service_check_spread and service_interleave_factor are
deprecated in Naemon since August of 2015 (Naemon commit 1b147702).

There is a very small amount of spam that makes me look twice at the
build output, when it shouldn't. Also, commented out code/config is a
bad pattern, because:
- they might no longer be supported
- the intent of still having them is undocumented
- it could make the reader a bit confused, "what are up with these?"

Signed-off-by: Carl Helmertz <chelmertz@op5.com>