cucumber/mock: Do not treat all fields ending with 's' as pluralized
The mocking library tries to be smart by assuming that everything that
ends with an 's' is plural instead of singular.

I tried to mock the alias field, but it did not show up on any
screenshots. When debugging the ORMDriverNative, the alias column was
apparently being fed an array, because the part of our mocking library
that is used from Cucumber was a bit optimistic.

I'm going ahead to leave a comment in the code. I'm all for better ideas
to solving this code, but the only thing I can think of is to parse the
ORM structure files (which are written in PHP) and feed the data to this
Ruby script but.. I don't want to rewrite that part of our structures to
make a one-off error test case work again. If you find this commit
message after reading that comment, we might have to look into it again
though.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>