Unit-of-measurement size now follows calculated font size
The UoM for perf data on Big Number widget was fixed to 50% of
font size. This made it look very different than it's non-perf data
counter part (i.e. using filter as content).

The font size for the perf data value is dynamically calculated and
depends on the string length of the value. It will look weird (for
some values) to have a dynamic size on the value and not the UoM so
I changed this into using the calculated fint size as well. This
forces inline styling though :(

The above make it look very similar (for short numbers) to the
"normal" Big Number widget value. I cannot quite understand why we
render them differently though (the fact that perf data could be
a very long number might be an explaination).

I also removed the space between the value and the UoM by removing
all html whitespace between them.

Signed-off-by: Joachim Ekstrand <jekstrand@op5.com>