backup: Pass $backup_directory to constructor and make it private
Tomas made a good point in an already merged pull request. I reasoned
that I had to make it public because Kohana always creates the
controller by itself and cannot pass arguments to the constructor (in
Kohanas current form).

Tomas realized that not passing any arguments can work in our favor,
because we can keep the hardcoded value as a default value for the
constructor argument, something that passed right by me when I wished
for DI and moved on. This looks much better though.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>