Use required attributes for form fields
Not all form fields yet, only text and number fields.

Since we now have optional values, the inverse is true, i.e. that the other
fields are required. Reflect this in the HTML so that users get direct feedback
before submitting to form in vain only to get an error message.

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>