js: Set timeout to 9 seconds for bundled "unhandled problems" icons
The issue MON-6782 says that a browser window with a listview doesn't
self heal if these steps are followed:
- open a listview in Monitor
- stop Naemon
- wait for a reload of the listview (set the "listview refresh rate" to
  10 before reproducing this step)
- an alert box is shown
- start Naemon again

Now, the listview should heal again, and it does (it was fixed after the
issue was logged).

There are other consequences though, the ajax requests triggered from
the "unhandled problems" icons never timed out, so while Naemon was
stopped and the listview was open, really many requests were opened and
never closed. An unhandled problem request is started every 10 seconds,
so we set the timeout of each request to 9 seconds to avoid overlapping
requests.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>