URLs with leading / should not be prefixed with /monitor/index.php
Dokuwiki links begin with '/dokuwiki' and are found at that path,
relative the web server's root folder.

We use url::redirect($url) to redirect the user to any link, but
url::redirect() wanted to prefix the $url argument with
"/monitor/index.php" (or /ninja/index.php in the open source
version). The reason for this is that we want to call
url::redirect("avail/index") and make Ninja still find the correct
path.

This patch makes sure that $url arguments that start with "/" are
treated as absolute paths (they are not prefixed with the common
prefix).

We also take the URL mangling logic out of url::redirect() and place
it in url::_redirect() in order to make it testable. url::redirect()
is still the public API and adds the actual exit() call and sends
the headers. This was verified manually by visiting
/monitor/index.php/cli, which was correctly redirected to "the tac".

A small side effect is that we remove the ability to call
url::redirect($url, "refresh"), after having looked through
/opt/monitor/op5 after that usage, ending up not finding anything.

This fixes MON-9648

Signed-off-by: Carl Helmertz <chelmertz@op5.com>