js: Do not add the same part of the URL when refreshing
This was easily reproducible in Nacoma:

1) Go to Manage -> Configure -> Hosts
2) Ctrl+R three times and you'll see a "page not found" message. The URI
keeps appending on each refresh, this is often seen if you move between
pages in Nacoma.

nacoma_base always replaced the first or the last "/" in the part of the
path being looked into. If _site_domain was "/monitor/", then the first
"/" would be removed, but the second "/" would be left on. Since one of
the parts ended with "/" and we later joined the array with "/", we
would get "monitor//index.php" pretty often.

Adding a leading slash effectively fixes the string so that it can be
matched and removed by the following substr() call.

I also pulled out the variable new_url both to document the intention,
and to have a better target for debugging in Chrome's dev tools.

This fixes MON-9651.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>