backup: Namespace HTML selectors to avoid clashes
There was a recent change in Ninja which made all JavaScript files
executed on every page. Such an approach requires a well thought out
strategy of namespacing, which we currently do not employ. This patch is
the result of such a mismatch, the backup views expect to be the sole
user of selectors such as #verify, .view, .restore and .delete. These
selectors clashes with the BSM view, but we'll begin making the changes
on this side of things.

This may be the cause of failing BSM builds in our CI environment.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>