test: Remove unused assertArraySubset()
It was added because we wanted to match user rights, but it ended up
never being used.

When building for EL7, which has access to a newer PHPUnit, the method
assertArraySubset() already exists as a static method, and it cannot be
overwritten by an instance method like we have here.

This fixes issue MON-9892.

Change-Id: I511aa6acf628c1d580237330c3708d696bd6509f
Signed-off-by: Carl Helmertz <helmertz@gmail.com>
(cherry picked from commit e9d3f2229e3f9b2345415f33000c7d268fb19058)