test: Remove unused assertArraySubset()
It was added because we wanted to match user rights, but it ended up
never being used.

When building for EL7, which has access to a newer PHPUnit, the method
assertArraySubset() already exists as a static method, and it cannot be
overwritten by an instance method like we have here.

Change-Id: I511aa6acf628c1d580237330c3708d696bd6509f
Signed-off-by: Carl Helmertz <helmertz@gmail.com>