orm/structure: Remove commented/dead code
There are some tables that exist in Livestatus but are not directly
exposed in Ninja's ORM, such as servicesbyhostgroup.

Although it could be a good hint of their existence, we remove them from
this location because 1) it is hard to find, and 2) the information may
be very dated and not useful for us now. That's the thing with commented
out (or unused) code - we have no idea if it's useful or not.

Change-Id: Id879d1c202b714fae8dd85a9ea7999d08af87101
Signed-off-by: Carl Helmertz <helmertz@gmail.com>