ORMObjectGenerator: Foreign keys are defined with arrays, not strings
When building for EL7, with a more strict PHP setting/version than what
is used when building for EL6, we notice some previously hidden
errors. In this case, we tried to refer to an array as a string.

The context is that we read data from structure.php in all of Ninja's
modules, when we build Ninja. Some of the fields in the structure are
foreign keys though, which led to this:

$ grep -m1 -A9 factory_from_array modules/monitoring/models/base/basecomment.php
	static public function factory_from_array($values, $export) {
		$obj = new static();
		$structure = array(
		'author' => 'string',
		'comment' => 'string',
		'entry_time' => 'time',
		'entry_type' => 'int',
		'expire_time' => 'time',
		'expires' => 'int',
		'host' => 'Array',

You see the "Array", right? :/

This fixes MON-9894.

Change-Id: Ia9f078aae431c5653f3c4c43b3edeb58e955c5e1
Signed-off-by: Carl Helmertz <helmertz@gmail.com>
(cherry picked from commit 7d41b1addc392bc68994722c4bc9c7a24ad33d9a)