sql: Delete recurring downtime (RD) objects if RD is deleted
When building on EL7 and executing tests against MariaDB, we noted
that foreign keys were not properly enforced previously, leading to
errors such as:

    Cannot delete or update a parent row: a foreign key constraint
    fails (`merlin_test`.`recurring_downtime_objects`, CONSTRAINT
    `recurring_downtime_objects_ibfk_1` FOREIGN KEY
    (`recurring_downtime_id`) REFERENCES `recurring_downtime` (`id`)) -
    DELETE FROM recurring_downtime WHERE comment = 'Recurring Downtime
    Test Schedule For hosts'

To avoid that, try to delete the previously existing foreign key, and
add a new one, with the addition to cascade the deletes. Cascading
deletes means that when a row in recurring_downtime is deleted, all rows
in recurring_downtime_objects that refer to the now-deleted
recurring_downtime, are also deleted.

Change-Id: Ia4c4c4b2e7bf9e0dedf154b4515a0e70f58b1be6
Signed-off-by: Carl Helmertz <helmertz@gmail.com>