test: migrate_auth.php should never miss any auth rights
We just added traps_view_all to the auth_groups.yml, the ACL in
load_mayi_rights.php and the canonical op5Authorization.

We missed adding it to the migration script, because nothing told us
that it was necessary. The tests that are included in this patch makes
sure that never happens again.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>