migrate_auth.php: service_view_all grants traps_view_all
The previous logic in Trapper only checked for service_view_all. When we
added the new right traps_view_all, we are (at least initially) looking
for "either of those rights". In the long term, traps_view_all is what
we should look for, but rights-wise, they are currently the same.

This is part of MON-9718.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>