mock.rb: Save temp data to relative path
Under EL7 services running in systemd have private namespaces for /tmp
This means that when we tell a script to save a file to /tmp the file will
be saved in /tmp, but when eg our webserver will look for the file it will
try to locate the file under /tmp/systemd-private-namespace....httpd...

So when we add our custom header to tell httpd what mock data to use it
doesn't find the file.

This fixes issue MON-9896.

Signed-off-by: Mattias Ryrlén <mattiasr@op5.com>
(cherry picked from commit 0fd8602101879f8a9ea103df3042c9ed97b0b266)