jquery-ui: 1.12.1 unzipped and placed in application
This patch is the whole content of the official, "quick
download"/"stable" link on jquery-ui's web page.

I will remove and move around some of the files in the following
commits, since we need to place js files in a special directory, and we
also want to skip unused files.

This is part of MON-9820

Signed-off-by: Carl Helmertz <chelmertz@op5.com>