about.js: Open about info in Lightbox, not fancybox
We are in the process of migrating away from Fancybox, since Fancybox
was a singleton and we needed more than one "layer" (in the depth plane)
of modal windows. Since we are upgrading to jQuery 3.1.1, the
deprecation of Fancybox quickly became a must issue, since Fancybox
relies on old(er) jQuery anyways.

This is part of MON-9820

Signed-off-by: Carl Helmertz <chelmertz@op5.com>