lsfilter: Always attach floating header
The listview's table layout always has its column names visible, even
when the user scrolls down. The code was written some time ago, with IE7
support in mind, and this patch simply removes that support since we
today are more interested in what's happening with >=IE11.

This is part of upgrading jQuery (MON-9820), where removing usage of
$.browser is needed.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>