Extract quicklink to Form_Model and own controller
The logic surrounding the quicklinks were embedded in common.js and was
not very generic. We try to approach this matter by shoehorning it into
the new Form_Model.

The catalyst for doing this refactoring is the imminent need for
removing the Fancybox library out of Ninja, since it both depends on an
older version of jQuery than the one we are updating to (3.1.1), but
also since the Fancybox is deprecated in favor of our own Lightbox which
can render more than one modal at a time.

This is part of MON-9820

Signed-off-by: Carl Helmertz <chelmertz@op5.com>