css: Adjust contextMenu in listview to jQuery 3.1.1
$.browser is deprecated in jQuery since 1.9.

The styles of the menu were removed together with the generic jQuery UI
stylesheets. I added a similar style to what we used before, but did not
adjust it to the separate skins because that's an overhead for an
already delayed deliverance of this issue.

This fixes MON-9838 and is part of MON-9820.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>