quicklink: Render quicklinks in a table
The previous design included a <li> that wrapped and aligned pretty
badly. Now we render the contents of the <td>s with a fixed set of
letters (sigh) to avoid getting a horizontal scrollbar, and also to
avoid specifying a fixed max-width of those specific <td> elements.

In the future, this should be a listview.

This fixes part of MON-9827

Signed-off-by: Carl Helmertz <chelmertz@op5.com>