lightbox: Remove unused remove_all()/topmost(), do not close on .cancel
form.js already attaches a click handler on the .cancel element of a
form-in-a-lightbox, which means that if we add another one in
LightboxManager.confirm(), we close two lightboxes at once, which was
annoying to track down. The funny thing is that the event handler for
the escape button only closed a single lightbox.

We also adjust the CSS a bit to compensate for missing styles regarding
button.cancel, which should be rendered as the secondary option.

To top all of this off, we also remove two unused methods:
- LightboxManager.remove_all(), and
- LightboxManager.topmost()

This fixes UX feedback in MON-9827

Signed-off-by: Carl Helmertz <chelmertz@op5.com>