Move upgrade_recurring_downtime() from Kohana to separate script
The build logs are spammed with HTML output during the part of the
installation that handles SQL migrations. That is because we expose the
scripts as part of Kohana so that they need to adapt to state (user
sessions) and hooks, both of which are unnecessary when being called in
the installation step. The side effects are not only unnecessary but
also triggers invalid template logic, which is what we need to get rid
of.

Instead, we move the script to the install_scripts folder, where they
reside outside of Kohana.

This solves part of MON-9764.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>