Move db migration (v13 -> v14) from Kohana to separate script
By having logic inside of Kohana, we need to handle template logic and
such. Since this script is only during the installation process, we are
outside of Kohana and do not want to bother with things like requests,
user sessions, Kohana hooks and the like -- because all those things
causes the installation to spit out lots of HTML during the installation
process, which is avoided by moving the logic out to separate scripts.

This fixes part of MON-9764.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>