log: Make ninja syslog with file=syslog in /etc/op5/log.yml
Dogfooding is a good thing. We should be able to use Logger in order to
track what Ninja is up to. Now you can, set the file property to
"syslog" for any log namespace found in /etc/op5/log.yml.

By just inserting a stupid if-case in the middle of the code, we are
able to keep the same logging interface that is used throughout the
product.

One comment in MON-7786 states that we should configure our syslog
handler to split out our messages into separate files. I don't believe
that is within the scope of MON-7786, but rather if/when we start to
syslog by default, i.e. change the actual default values. For now,
nothing will change except if you know of this change and manipulate the
configuration files yourself.

This fixes MON-7786.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>