ci: Allow cuke tests on el6, skip only on el7
Some cuke tests fail on el6 and has done so for a long time. Many other
tests fail only on el7. We must keep apart these tests that fail on
either el6 or el7, in order to still run as many tests as possible on
each architecture.

This patch introduces @unreliable_el7 as a complement to @unreliable

Signed-off-by: Carl Helmertz <chelmertz@op5.com>