Add ORM field type classes
This makes ORM field types more extensible, in addition it grants support for
more complex types beyond the implicit Livestatus relation style while
remaining backwards compatible in order to not break other repositories or
existing structures.

In order to provide a consistent interface to the new ORM Set type, which is
stored as a query, a possibility to store empty queries was required.
"[<object>] all" already existed and was reflected in the pool classes all()
method, however pool classes also provided a none() method as its complement.
So this has been reflected in the lsfilter grammar as well, you can now run
"[<object>] none" for an empty set. So that empty sets can be stored using the
ORM Set type in a generic manner.

This is required in order to use the ORM for all types of configurable objects
in a generic way, such as what is currently handled by Nacoma.

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>