orm: Expose the ORM structure through Kohana
When we mock data, the incoming HTTP request contains a header field
that contains a path to a JSON stored at the Monitor machine. Ninja
then, through a hook, reads that JSON and replaces/hijacks the backend
for every table mentioned in the JSON file.

The thing is, when we replace the real backend with the mocked backend,
we do not properly check that the mocked fields exists. For example, we
allow mocking of "rumplebutt" on the table "hosts", even if that table
does not contain that column/property in its real backend.

In order to validate the mocked data, we must expose the ORM data
somehow. JSON is a good way of making sure both PHP and Ruby understands
the data. We could either export the ORM through the HTTP API, or
through Ninja's controller. I chose the latter because it can
conveniently be called through CLI, thus it is readily available for
consumption through Ruby. The downside of exposing stuff through the
HTTP API is that consumers treat the interface as consistent. I think
promoting endpoints after having existed as Kohana controllers for a
while, might be a very good thing to exercise.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>