Use __call parameters appropriately
Since __call is automatically invoked if no accessable method is found with the
requested method name and an array of parameters passed it can NEVER recieve an
int (instance_id), as the second parameter.

This previously failed silently as the instance_id is only ever used in
rendering to separate one widget of the same type from another in dashboards.

But with new stricter ORM types it could not set an array (parameters) as instance_id,
when testing external widgets in nagvis.

This is part of MON-10173

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>