Changed details to go with API definition and made design more dynamic.
Since the API definition changed we also made changes to our mocked data, to
make it easier to connect API and UI in the future.
Design is also made more dynamic. HTML will be generated from existing data.
If you need to add/delete steps in exportsave dialog, you only have to do
changes in controllers/exportsave.php

This is a part of MON-10822.

Signed-off-by: Elin Linder <elinder@op5.com>