Refactored code and more basic design.
In this commit we've gone back to the Monitor design. Nothing new added, but
we're mostly using previous classes.
Code has been refactored, so that we use one function, instead of 3, to render
design. Also added functionality to show more and better info in lightbox,
such as a title bar and which step we're currently at.

This is part of MON-10822

Signed-off-by: Elin Linder <elinder@op5.com>