ninja: Display of contact, contacgroups and comments on extinfo page.
After LMD integration to monitor, the extinfo pages for hosts
or services were not displaying properly for contact, contactgroup
or comments. The reason being livestatus used to accept even incorrect
comments requests which LMD discards. And in case of contacts or
contactgroups a check is added to load only if something is there.
Otherwise it was going to a permanent loading state even if there is
 no data to display.
Because all three (contact, contactgroups and comments) are widgets
to the extinfo page. Render the widgets only if there is some
data else don't render and display a permanent "Loading" text.

This fix addresses the issue#11243.

MON-11243

Signed-off-by: Sidharta Bhapur <sbhapur@op5.com>