LDAP use empty to check password and add unit test
This commit adds a test to ensure that an empty password results in a
failed login.

Futhermore we use the empty() function to test if password is empty/not
set, rather than checking manually.

Note that ldap_bind attemts an annonmous login if the password is not
set, and therefore returns true, if that is allowed. But in that case
the user would be logged in to OP5 as the actual OP5 user, which is bad.

Signed-off-by: Jacob Hansen <jhansen@op5.com>